pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A django-ninja backend template"
dependencies = [
    "uwsgi; platform_system != 'Windows'",
    "pick",
    "django-ninja",
    "django-extensions",
    "django>=3.0",
    "arrow",
    "PyJWT",
    "django-cors-headers",
    "mysqlclient",
    "cryptography",
    "requests",
    "redis==4.6.0",
    "loguru",
    "django-constance",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.optional-dependencies]
dev = [
    "uv",
    "ruff",
    "rich",
    "ipython",
    "types-redis~=4.6.0",
    "django-print-sql",
]

[tool.setuptools.packages.find]
include = ["backend", "backend.*"]

[tool.ruff]
exclude = [".venv", "venv", "typings", "backend/apps/*/migrations"]
line-length = 160
target-version = "py310"

[tool.ruff.lint]
extend-select = [
    "A",    # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
    "B006", #https://docs.astral.sh/ruff/rules/mutable-argument-default/
    "I",    #https://docs.astral.sh/ruff/rules/#isort-i
    "INP",  #https://docs.astral.sh/ruff/rules/#flake8-no-pep420-inp
    "UP",   #https://docs.astral.sh/ruff/rules/#pyupgrade-up
]
ignore = [
    "F401", # https://docs.astral.sh/ruff/rules/unused-import/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
]

[tool.uv.pip]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.rye.sources]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
[[tool.rye.sources]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
[[tool.rye.sources]]
name = "pypi"
url = "https://pypi.org/simple/"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "rich",
    "ipython",
    "types-redis~=4.6.0",
    "ruff",
    "uv",
    "django-print-sql",
]

[tool.rye.scripts]
start = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
tail = "python watch_log.py"
shell = """python manage.py shell_plus \
--ipython \
--print-sql \
--dont-load auth \
--dont-load admin \
--dont-load contenttypes \
--dont-load sessions"""
